FCN-244 Consolidate registry modules
