+++ packages/nxtcm-rosa-hcp-wizard/src/Footer/RosaHcpWizardFooter.spec-helpers.tsx
@@ -157,7 +157,6 @@ export const RosaHcpWizardValidationMount: React.FC<RosaHcpWizardValidationMount
                   awsBillingAccounts={awsBilling}
                   regions={regionsProps}
                   roles={rolesProps}
-                  vpcList={vpcListProps}
                 />
               </WizardStep>,
               <WizardStep
