+++ packages/nxtcm-rosa-hcp-wizard/src/syncFieldsOnSourceChange.ts
@@ -1,12 +1,17 @@
 import * as yup from 'yup';
 import type { FieldPath, FieldPathValue, UseFormSetValue } from 'react-hook-form';
 
-import { buildFormSetValueOptions, type FormSetValueOptions } from './formSetValueOptions';
 import type { ROSAHCPCluster } from './types';
 import { clusterValidationSchema } from './yupSchemas';
 import type { WizardFieldSyncOnChange, WizardFormFieldName } from './yupSchemas/types';
 
 type FormPath = FieldPath<Partial<ROSAHCPCluster>>;
+
+type FormSetValueOptions = {
+  shouldDirty?: boolean;
+  shouldTouch?: boolean;
+  shouldValidate?: boolean;
+};
 
 export type SyncFieldsOnSourceChangeOptions = FormSetValueOptions & {
   /** When true, only `clear` runs — used on initial mount to drop stale inactive fields without overwriting hydrated values. */
@@ -30,7 +35,11 @@ export function syncFieldsOnSourceChange(
     return;
   }
 
-  const setOpts = buildFormSetValueOptions(options);
+  const setOpts = {
+    shouldDirty: options.shouldDirty ?? true,
+    shouldTouch: options.shouldTouch ?? false,
+    shouldValidate: options.shouldValidate ?? false,
+  };
 
   for (const name of branch.clear ?? []) {
     setValue(
