--- packages/nxtcm-rosa-hcp-wizard/src/yupSchemas/listWizardFieldMetaChangeSourceFields.test.ts
@@ -1,8 +1,0 @@
-import { listWizardFieldMetaChangeSourceFields } from './listWizardFieldMetaChangeSourceFields';
-
-describe('listWizardFieldMetaChangeSourceFields', () => {
-  it('includes fields that declare derivedFieldsSyncOnChange in Yup meta', () => {
-    expect(listWizardFieldMetaChangeSourceFields()).toContain('installer_role_arn');
-    expect(listWizardFieldMetaChangeSourceFields()).toContain('selected_vpc');
-  });
-});
